Use BigInteger in Level

+++ base/src/main/java/org/arend/term/abs/AbstractExpressionVisitor.java
@@ -14,8 +14,8 @@ public interface AbstractExpressionVisitor<P, R> {
   R visitThis(@Nullable Object data, P params);
   R visitLam(@Nullable Object data, @NotNull Collection<? extends Abstract.LamParameter> parameters, /* @NotNull */ @Nullable Abstract.Expression body, P params);
   R visitPi(@Nullable Object data, @NotNull Collection<? extends Abstract.Parameter> parameters, /* @NotNull */ @Nullable Abstract.Expression codomain, P params);
-  R visitUniverse(@Nullable Object data, @Nullable Integer pLevelNum, @Nullable Integer hLevelNum, @Nullable Abstract.LevelExpression pLevel, P params);
-  R visitCatUniverse(@Nullable Object data, @Nullable Integer pLevelNum, @Nullable Abstract.LevelExpression pLevel, P params);
+  R visitUniverse(@Nullable Object data, @Nullable BigInteger pLevelNum, @Nullable Integer hLevelNum, @Nullable Abstract.LevelExpression pLevel, P params);
+  R visitCatUniverse(@Nullable Object data, @Nullable BigInteger pLevelNum, @Nullable Abstract.LevelExpression pLevel, P params);
   R visitSortUniverse(@Nullable Object data, P params);
   R visitApplyHole(@Nullable Object data, P params);
   R visitInferHole(@Nullable Object data, P params);
